add pytests, set flag to false for adding parameters in hybrid, add if for flag in apply
